+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Enable declaration of subtypes through `withSubtypeResolver(SubtypeResolver)` on `forTypesInGeneral()` (#24)
+
+### Changed
+- Move custom definitions and type attribute overrides into `forTypesInGeneral()` (while preserving delegate setters on config builder)
 
 ## [4.3.0] - 2020-02-28
 ### Changed